api/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-cloud-bigquery==3.40.0
python-dotenv==1.2.1
